requirements/requirements_doc.txt
#PyMechanical
ansys-mechanical-core[doc]==0.11.14
ansys-mechanical-core[rpc]==0.11.14